requirements.txt
scipy
plotly.express
pandas 
numpy
matplotlib
seaborn
sklearn
streamlit